add vhive_fc_vmid resolver
